PythonServer/train_autoencoder.py
"""
SentinelVR — Treinamento do Autoencoder de Detecção de Anomalias
=================================================================
Script para treinar o autoencoder convolucional com frames normais capturados
das câmeras de vigilância.

Uso:
    python train_autoencoder.py --data_dir ./data/normal_frames --epochs 50 --output model.pth

Estrutura esperada de dados:
    data/normal_frames/
        frame_0001.jpg
        frame_0002.jpg
        ...
"""

import argparse
import logging
import os
import torch
import torch.nn as nn
from torch.utils.data import DataLoader, Dataset
from torchvision import transforms
from PIL import Image

# Importar arquitetura do servidor
from anomaly_server import ConvAutoencoder

logging.basicConfig(level=logging.INFO, format="%(asctime)s [%(levelname)s] %(message)s")
logger = logging.getLogger("SentinelVR.Train")


# ─────────────────────────────────────────────
# Dataset
# ─────────────────────────────────────────────

class SurveillanceFrameDataset(Dataset):
    """
    Dataset de frames normais de câmeras de vigilância para treino do autoencoder.
    """

    SUPPORTED_EXTENSIONS = {".jpg", ".jpeg", ".png", ".bmp"}

    def __init__(self, data_dir: str, image_size: int = 224):
        self.data_dir = data_dir
        self.transform = transforms.Compose([
            transforms.Resize((image_size, image_size)),
            transforms.ToTensor(),
        ])
        # TODO: Carregar lista de arquivos de imagem do data_dir
        self.image_paths = []
        self._load_image_paths()

    def _load_image_paths(self):
        """Carrega caminhos de todas as imagens suportadas no diretório."""
        # TODO: Iterar sobre data_dir e filtrar por SUPPORTED_EXTENSIONS
        pass

    def __len__(self):
        return len(self.image_paths)

    def __getitem__(self, idx):
        # TODO: Carregar imagem, aplicar transform, retornar tensor
        pass


# ─────────────────────────────────────────────
# Treinamento
# ─────────────────────────────────────────────

def train(model: ConvAutoencoder, dataloader: DataLoader, epochs: int,
          learning_rate: float, device: str, output_path: str):
    """
    Treina o autoencoder usando reconstrução MSE como função de perda.
    Salva checkpoint ao final de cada época e o modelo final em output_path.
    """
    optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
    criterion = nn.MSELoss()

    model.train()
    for epoch in range(epochs):
        total_loss = 0.0
        num_batches = 0

        for batch in dataloader:
            # TODO:
            # 1. Mover batch para device
            # 2. Forward pass: output = model(batch)
            # 3. Calcular loss = criterion(output, batch)
            # 4. Backward pass e optimizer.step()
            # 5. Acumular total_loss
            pass

        avg_loss = total_loss / max(num_batches, 1)
        logger.info(f"Época [{epoch+1}/{epochs}] Loss: {avg_loss:.6f}")

        # TODO: Salvar checkpoint a cada 10 épocas

    # TODO: Salvar modelo final em output_path
    logger.info(f"Treinamento concluído. Modelo salvo em: {output_path}")


# ─────────────────────────────────────────────
# Avaliação
# ─────────────────────────────────────────────

def evaluate_threshold(model: ConvAutoencoder, normal_dir: str,
                        anomaly_dir: str, device: str):
    """
    Avalia o limiar ideal de detecção comparando frames normais e anômalos.
    Calcula métricas: precisão, recall, F1-score para diferentes thresholds.
    """
    # TODO: Implementar avaliação com exemplos de anomalias rotuladas
    pass


# ─────────────────────────────────────────────
# Main
# ─────────────────────────────────────────────

def main():
    parser = argparse.ArgumentParser(description="Treinar autoencoder SentinelVR")
    parser.add_argument("--data_dir", required=True, help="Diretório com frames normais de treinamento")
    parser.add_argument("--epochs", type=int, default=50, help="Número de épocas (default: 50)")
    parser.add_argument("--batch_size", type=int, default=32, help="Tamanho do batch (default: 32)")
    parser.add_argument("--lr", type=float, default=1e-3, help="Learning rate (default: 0.001)")
    parser.add_argument("--output", default="sentinel_model.pth", help="Caminho do modelo salvo")
    parser.add_argument("--image_size", type=int, default=224, help="Tamanho das imagens (default: 224)")
    args = parser.parse_args()

    device = "cuda" if torch.cuda.is_available() else "cpu"
    logger.info(f"Dispositivo: {device}")

    # TODO: Criar dataset e dataloader
    # TODO: Instanciar e treinar modelo
    # TODO: Chamar train() com os parâmetros corretos

    logger.info("Script de treinamento pronto. Implemente os TODOs para treinar o modelo.")


if __name__ == "__main__":
    main()
